Fix the number of removed functions in change report
Right now, the number of removed functions that is shown in the change
report is the total number of removed functions.

But it should be the *net* number of removed functions.  That is, the
total number of removed functions minus the number of removed
functions that were filtered out.

This patch fixes that.  A one liner.

	* src/abg-suppression.cc (corpus_diff::report): Show the net
	number of removed functions, not the total number of the removed
	functions.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>