Don't require all version symbol sections to present
When we need to get the version of a defined symbol, only the
SHT_GNU_versym and SHT_GNU_verdef sections are necessary.
SHT_GNU_verneed is not necessary, for instance.

So do not require that all of the three version-related sections to be
present when we want to get some symbol version information.

Otherwise, we just don't get the version of a defined symbol when the
SHT_GNU_verneed section is not present.  I stumbled upon this while
looking the abidw's output of ld-2.17.so from
glibc-2.17-79.el7_1.x86_64.  The _rtld_global_ro variable's symbol was
being seen as having no symbol version.  In reality it has the
GLIBC_PRIVATE version, but because the binary lacks a SHT_GNU_verneed
section, we were not getting the symbols version information.

I am not adding that library to the test suite because it's too big.
But at least this change doesn't break the existing regression test
suite.

	* src/abg-dwarf-reader.cc (get_symbol_versionning_sections): Allow
	returning just some of the three version-related section, not
	necessarily all of them.  Adjust comment.
	(get_version_for_symbol): Be ready to not necessarily having the
	three version-related sections available.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>