Consider a method_decl as always being a member decl
The is_member_decl() function considers that a decl is a member decl
only if it's at class scope.

In preparation for the support of incremental building of method
decls, a method decl, even before being added to its final class, must
always be considered as a member decl.  This allows for instance, the
proper pretty printing of the method decl even before it's added to
its class.

This patch prepares the upcoming support of incremental building of
method decls by making is_member_decl() work on a method_decl that
hasn't yet been added to a class, so that is not, strictly-speaking,
at class scope yet.

	* src/abg-ir.cc (is_member_decl): Consider a method decl as always
	being a member decl.
	(is_member_function): Use is_member_decl.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>