Factorize out parsing of integral types
Until now, integral_type has been just a "private" helper type used in
the building of a type_decl.

Now it's going to be helpful in determining the name of an integral
type, directly from DIEs -- without having to build a type_decl.

This patch factorizes out the parsing and building of an integral_type
by introducing a src/abg-ir-priv.h file that is meant to declare
interfaces that are going to be private to libabigail.so but usable by
all of its modules.

	* src/abg-ir-priv.h: New file.
	* src/Makefile.am: Add abg-ir-priv.h to the build system.
	* src/abg-ir.cc: Include the new abg-ir-priv.h header file.
	(class_integral_type): Move this type
	declaration to the new abg-ir-priv.h header.
	(integral_type::modifiers_type): Make this non-static.
	(parse_integral_type): This new overload is a factorized out of
	...
	(integral_type::integral_type): ... here.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>